Use a percentage instead of fraction for the remote connections percentage CLI option. (#1682)
* Factorise metrics code for KeyValueStorage database

- introduce `MonitorableKeyValueStorage`
- factorise code
- remove metrics instanciation in `RocksDbKeyValueStorage` and `ColumnarRocksDbKeyValueStorage`

* Rename class

* Use percentage instead of fraction

* add tests

* fix PR discussion

- introduce `PercentageConverter` and `FractionConverter` in `tech.pegasys.pantheon.cli.converter`
- use custom error message for `PercentageConversionException
`

* fix PR discussion

- return primitive types instead of objects for converters

* fix tests

+++ ethereum/p2p/src/main/java/tech/pegasys/pantheon/ethereum/p2p/config/RlpxConfiguration.java
@@ -22,7 +22,7 @@ import java.util.List;
 import java.util.Objects;
 
 public class RlpxConfiguration {
-  public static final Double DEFAULT_FRACTION_REMOTE_CONNECTIONS_ALLOWED = 0.5;
+  public static final float DEFAULT_FRACTION_REMOTE_CONNECTIONS_ALLOWED = 0.5f;
   private String clientId = "TestClient/1.0.0";
   private String bindHost = "0.0.0.0";
   private int bindPort = 30303;
